requirements.txt
# Requirements for the SQLite dictionary project
boto3>=1.34.0
botocore>=1.34.0
requests>=2.31.0
rich>=13.0.0
python-dotenv>=1.0.0
Flask>=3.0.0

# Optional: For database management and visualization
sqlite-utils>=3.35.0  # Command-line tool for SQLite operations
# Optional: For OpenAI API interactions
annotated-types>=0.7.0
anyio>=4.11.0
certifi>=2025.10.5
distro>=1.9.0
h11>=0.16.0
httpcore>=1.0.9
httpx>=0.28.1
idna>=3.10
jiter>=0.11.0
openai>=2.3.0
pydantic>=2.12.0
pydantic_core>=2.41.1
sniffio>=1.3.1
tqdm>=4.67.1
typing-inspection>=0.4.2
typing_extensions>=4.15.0

celery>=5.3.6
redis>=5.0.4
tenacity>=8.2.0